* Documentation of UI State and Domain States related to Splits
* Fix manual split going crazy with shares values when splitting between
  more than 2 shareholders
